Code Refinement

+++ src/Application/HouseBrokerApplication.Application/Interface/ICommissionService.cs
@@ -8,6 +8,6 @@ namespace HouseBrokerApplication.Application.Interface
 {
     public interface ICommissionService
     {
-        decimal CalculateCommission(decimal price);
+        double CalculateCommission(double price);
     }
 }
