+++ src/Domain/HouseBrokerApplication.Domain/Entities/CommissionRate.cs
@@ -9,8 +9,8 @@ namespace HouseBrokerApplication.Domain.Entities
     public class CommissionRate
     {
         public int Id { get; set; }
-        public decimal MinPrice { get; set; }
-        public decimal MaxPrice { get; set; }
+        public double MinPrice { get; set; }
+        public double MaxPrice { get; set; }
         public double Percentage { get; set; }
     }
 }
